+++ docs/en/changes/changes.md
@@ -63,6 +63,7 @@
 * Add Service Hierarchy auto matching layer relationships (upper -> lower) as following:
   - RABBITMQ -> K8S_SERVICE
   - VIRTUAL_MQ -> RABBITMQ
+* Make query `readMetricValue` always return the average value of the duration.
 
 #### UI